+++ e-httpRequest/CHANGELOG.md
@@ -6,6 +6,7 @@ v1.1.0
 * Support for HTTPS
 * Support for dynamic configuration (only URL of the request and request body for raw mode)
 * Content type may be customized in the file mode
+* Support for multiple request bodies with form formParams (configured via RDF dynamic configuration)
 
 v1.0.0
 ---